+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/ProductQueryBuilder.php
@@ -16,23 +16,21 @@ use Pim\Bundle\CatalogBundle\Doctrine\ProductQueryBuilderInterface;
  */
 class ProductQueryBuilder implements ProductQueryBuilderInterface
 {
-    /**
-     * QueryBuilder
-     * @var QueryBuilder
-     */
+    /** @var QueryBuilder */
     protected $qb;
 
-    /**
-     * Locale code
-     * @var string
-     */
-    protected $locale;
+    /** @var CatalogContext */
+    protected $context;
 
     /**
-     * Scope code
-     * @var string
+     * Constructor
+     *
+     * @param CatalogContext $catalogContext
      */
-    protected $scope;
+    public function __construct(CatalogContext $catalogContext)
+    {
+        $this->context = $catalogContext;
+    }
 
     /**
      * Get query builder
@@ -63,17 +61,7 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
      */
     public function getLocale()
     {
-        return $this->locale;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLocale($code)
-    {
-        $this->locale = $code;
-
-        return $this;
+        return $this->context->getLocaleCode();;
     }
 
     /**
@@ -81,17 +69,7 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
      */
     public function getScope()
     {
-        return $this->scope;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setScope($code)
-    {
-        $this->scope = $code;
-
-        return $this;
+        return $this->context->getScopeCode();
     }
 
     /**
@@ -114,14 +92,14 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
             $filterClass = 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\BaseFilter';
         }
 
-        $filter = new $filterClass($this->qb, $this->locale, $this->scope);
+        $filter = new $filterClass($this->qb, $this->getLocale(), $this->getScope());
         $filter->addAttributeFilter($attribute, $operator, $value);
 
         return $this;
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc}
      */
     public function addFieldFilter($field, $operator, $value)
     {
@@ -139,7 +117,7 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
             $filterClass = 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\BaseFilter';
         }
 
-        $filter = new $filterClass($this->qb, $this->locale, $this->scope);
+        $filter = new $filterClass($this->qb, $this->getLocale(), $this->getScope());
         $filter->addFieldFilter($field, $operator, $value);
 
         return $this;
@@ -159,7 +137,7 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
             $sorterClass = 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Sorter\BaseSorter';
         }
 
-        $sorter = new $sorterClass($this->qb, $this->locale, $this->scope);
+        $sorter = new $sorterClass($this->qb, $this->getLocale(), $this->getScope());
         $sorter->addAttributeSorter($attribute, $direction);
 
         return $this;
@@ -173,7 +151,7 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
         // TODO : need to be refactor to properly handle extra context parameters
         if (strpos($field, 'in_group_') !== false) {
             $sorterClass = 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Sorter\InGroupSorter';
-            $sorter = new $sorterClass($this->qb, $this->locale, $this->scope);
+            $sorter = new $sorterClass($this->qb, $this->getLocale(), $this->getScope());
             $sorter->addFieldSorter($field, $direction);
 
             return $this;
@@ -190,7 +168,7 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
             $sorterClass = 'Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Sorter\BaseSorter';
         }
 
-        $sorter = new $sorterClass($this->qb, $this->locale, $this->scope);
+        $sorter = new $sorterClass($this->qb, $this->getLocale(), $this->getScope());
         $sorter->addFieldSorter($field, $direction);
 
         return $this;